MA-10434: add secure playback feature for 2 input buffers.
Reviewed-on: http://androidsource.ap.freescale.net/project/3303
Reviewed-on: http://androidsource.ap.freescale.net/project/5465

+++ media/libmediaplayerservice/nuplayer/NuPlayerRenderer.cpp
@@ -1335,8 +1335,8 @@ void NuPlayer::Renderer::onDrainVideoQueue() {
         tooLate = (mVideoLateByUs > 40000);
 
         if (tooLate) {
-            ALOGV("video late by %lld us (%.2f secs)",
-                 (long long)mVideoLateByUs, mVideoLateByUs / 1E6);
+            ALOGV("video late by %" PRId64 " us (%.2f secs)",
+                 mVideoLateByUs, mVideoLateByUs / 1E6);
         } else {
             int64_t mediaUs = 0;
             mMediaClock->getMediaTime(realTimeUs, &mediaUs);
